+++ prawcore/auth.py
@@ -200,15 +200,34 @@ class Authorizer(BaseAuthorizer):
 
     AUTHENTICATOR_CLASS = BaseAuthenticator
 
-    def __init__(self, authenticator, refresh_token=None):
+    def __init__(
+        self,
+        authenticator,
+        *,
+        post_refresh_callback=None,
+        pre_refresh_callback=None,
+        refresh_token=None,
+    ):
         """Represent a single authorization to Reddit's API.
 
         :param authenticator: An instance of a subclass of :class:`BaseAuthenticator`.
+        :param post_refresh_callback: (Optional) When a single-argument
+            function is passed, the function will be called prior to refreshing
+            the access and refresh tokens. The argument to the callback is the
+            :class:`Authorizer` instance. This callback can be used to inspect
+            and modify the attributes of the :class:`Authorizer`.
+        :param pre_refresh_callback: (Optional) When a single-argument function
+            is passed, the function will be called after refreshing the access
+            and refresh tokens. The argument to the callback is the
+            :class:`Authorizer` instance. This callback can be used to inspect
+            and modify the attributes of the :class:`Authorizer`.
         :param refresh_token: (Optional) Enables the ability to refresh the
             authorization.
 
         """
         super(Authorizer, self).__init__(authenticator)
+        self._post_refresh_callback = post_refresh_callback
+        self._pre_refresh_callback = pre_refresh_callback
         self.refresh_token = refresh_token
 
     def authorize(self, code):
@@ -228,11 +247,15 @@ class Authorizer(BaseAuthorizer):
 
     def refresh(self):
         """Obtain a new access token from the refresh_token."""
+        if self._pre_refresh_callback:
+            self._pre_refresh_callback(self)
         if self.refresh_token is None:
             raise InvalidInvocation("refresh token not provided")
         self._request_token(
             grant_type="refresh_token", refresh_token=self.refresh_token
         )
+        if self._post_refresh_callback:
+            self._post_refresh_callback(self)
 
     def revoke(self, only_access=False):
         """Revoke the current Authorization.